refactor(accumulator-updater): refactor ixs & state into modules
