+++ accumulator_updater/programs/accumulator_updater/src/instructions/mod.rs
@@ -1,0 +1,7 @@
+pub use {
+    create_inputs::*,
+    update_inputs::*,
+};
+
+mod create_inputs;
+mod update_inputs;
