+++ accumulator_updater/programs/accumulator_updater/src/state/mod.rs
@@ -1,0 +1,7 @@
+pub use {
+    accumulator_input::*,
+    whitelist::*,
+};
+
+mod accumulator_input;
+mod whitelist;
